requirements.txt
Django==1.11.18
django-cors-headers==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.1
